refactor(chart): split types.ts into focused modules
Closes #26.

types.ts had grown to 728 lines holding 50+ exported symbols across
fundamental data, theme, drawing variants (12), series config, panes,
chart options, events, integration overlays, and the ChartInstance
interface. Splitting into focused modules under core/types/ improves
navigation and reduces the blast radius of future edits.

types.ts is now a re-export barrel, so all existing imports
(`from "../core/types"` and the package barrel) continue to work
without changes.
