+++ packages/chart/src/core/types.ts
@@ -1,728 +1,17 @@
 /**
  * Core type definitions for @trendcraft/chart
+ *
+ * This file re-exports from focused modules under `./types/`. Consumers
+ * should keep importing from `"../core/types"` (or via the package barrel
+ * `@trendcraft/chart`) — the split is internal.
  */
 
-// ============================================
-// Fundamental Types
-// ============================================
-
-/** Epoch milliseconds — matches trendcraft convention */
-export type TimeValue = number;
-
-/** Generic data point (compatible with trendcraft Series<T>) */
-export type DataPoint<T = number | null> = {
-  time: TimeValue;
-  value: T;
-};
-
-/** OHLCV candle data */
-export type CandleData = {
-  time: TimeValue;
-  open: number;
-  high: number;
-  low: number;
-  close: number;
-  volume: number;
-};
-
-// ============================================
-// Theme Types
-// ============================================
-
-export type ThemeColors = {
-  background: string;
-  text: string;
-  textSecondary: string;
-  grid: string;
-  border: string;
-  crosshair: string;
-  upColor: string;
-  downColor: string;
-  upWick: string;
-  downWick: string;
-  volumeUp: string;
-  volumeDown: string;
-};
-
-export const DARK_THEME: ThemeColors = {
-  background: "#131722",
-  text: "#d1d4dc",
-  textSecondary: "#787b86",
-  grid: "#1e222d",
-  border: "#2a2e39",
-  crosshair: "#758696",
-  upColor: "#26a69a",
-  downColor: "#ef5350",
-  upWick: "#26a69a",
-  downWick: "#ef5350",
-  volumeUp: "rgba(38,166,154,0.3)",
-  volumeDown: "rgba(239,83,80,0.3)",
-};
-
-export const LIGHT_THEME: ThemeColors = {
-  background: "#ffffff",
-  text: "#131722",
-  textSecondary: "#787b86",
-  grid: "#f0f3fa",
-  border: "#e0e3eb",
-  crosshair: "#9598a1",
-  upColor: "#26a69a",
-  downColor: "#ef5350",
-  upWick: "#26a69a",
-  downWick: "#ef5350",
-  volumeUp: "rgba(38,166,154,0.3)",
-  volumeDown: "rgba(239,83,80,0.3)",
-};
-
-// ============================================
-// Chart Configuration
-// ============================================
-
-export type ChartOptions = {
-  /** Chart width in pixels (default: container width) */
-  width?: number;
-  /** Chart height in pixels (default: 400) */
-  height?: number;
-  /** Color theme */
-  theme?: "dark" | "light" | ThemeColors;
-  /** Device pixel ratio (default: window.devicePixelRatio) */
-  pixelRatio?: number;
-  /** Right margin for price axis in pixels (default: 60) */
-  priceAxisWidth?: number;
-  /** Bottom margin for time axis in pixels (default: 24) */
-  timeAxisHeight?: number;
-  /** Font family (default: system) */
-  fontFamily?: string;
-  /** Font size in pixels (default: 11) */
-  fontSize?: number;
-  /** Custom price formatter (default: auto-precision) */
-  priceFormatter?: (price: number) => string;
-  /** Custom time formatter (default: smart date/time) */
-  timeFormatter?: (time: number) => string;
-  /** Watermark text displayed in chart background */
-  watermark?: string;
-  /** Show legend widget (default: true) */
-  legend?: boolean;
-  /** Show volume pane (default: true) */
-  volume?: boolean;
-  /**
-   * Show a colored last-value pill on the right price axis for every labeled
-   * series (including each channel of multi-channel series like Bollinger
-   * Bands / MACD) and for the volume pane. Default: false. The candle
-   * current-price badge is always shown regardless of this setting.
-   */
-  showSeriesBadges?: boolean;
-  /**
-   * Which "last value" to show in series badges:
-   * - `"absolute"` (default): the last non-null value in the underlying data
-   *   array — i.e. the live / latest known value, even if the viewport has
-   *   scrolled back.
-   * - `"visible"`: the last non-null value within the current visible range
-   *   — what the user sees at the right edge of the chart.
-   */
-  seriesBadgeMode?: "absolute" | "visible";
-  /** Scroll/pan sensitivity multiplier (default: 0.3, lower = slower) */
-  scrollSensitivity?: number;
-  /** Base chart type for price data (default: 'candlestick') */
-  chartType?: ChartType;
-  /** Custom info overlay formatter. Return HTML string or null to use default. */
-  formatInfoOverlay?: (data: InfoOverlayData) => string | null;
-  /** Animation duration in ms for range transitions (default: 300, 0 to disable) */
-  animationDuration?: number;
-  /** Locale strings for i18n (partial override of defaults) */
-  locale?: Partial<import("./i18n").ChartLocale>;
-  /** Maximum number of candles to retain in live mode. Older candles are trimmed when exceeded. */
-  maxCandles?: number;
-  /** Crosshair behavior */
-  crosshair?: CrosshairOptions;
-  /** Keyboard shortcut overrides. Pass `false` to disable all shortcuts (including Escape cancel). */
-  hotkeys?: Partial<Record<string, HotkeyAction>> | false;
-  /** Interaction behavior */
-  interaction?: InteractionOptions;
-  /** Time scale behavior (session gaps, etc.) */
-  timeScale?: TimeScaleOptions;
-};
-
-/**
- * Options for the time scale. Currently controls whether visual gaps are
- * inserted between trading sessions when the data is intraday.
- */
-export type TimeScaleOptions = {
-  /**
-   * Insert visual gaps between trading sessions (e.g. overnight breaks in
-   * intraday data) so the x-axis shows empty space rather than compressing
-   * non-trading hours to a zero-width seam.
-   *
-   * - `false` / omitted (default): index-based layout with no gaps (legacy).
-   * - `true`: auto-detect intraday (median bar interval < 6h) and insert a
-   *   gap wherever consecutive bars have a time delta significantly larger
-   *   than the median (e.g. overnight, weekends). TZ-agnostic.
-   * - object form: fine-grained control.
-   */
-  sessionGaps?: boolean | SessionGapsOptions;
-};
-
-export type SessionGapsOptions = {
-  /**
-   * Detection mode:
-   * - `"timeGap"` (default): insert a gap whenever the time delta between
-   *   consecutive bars exceeds `gapThresholdMs` (or 2 × median when unset).
-   *   Works correctly regardless of data timezone.
-   * - `"dayBoundary"`: insert a gap whenever the UTC calendar day changes
-   *   between consecutive bars. Useful for 24×7 data (crypto / FX) where
-   *   time deltas are uniform but you still want a daily divider.
-   * - `"off"`: disabled.
-   */
-  mode?: "timeGap" | "dayBoundary" | "off";
-  /** Gap size in bar-widths (default: 1.0). */
-  sizeBars?: number;
-  /** Only enable if median bar interval ≤ this many ms (default: 6h). */
-  intradayThresholdMs?: number;
-  /** For `"timeGap"` mode: the minimum delta to count as a session gap (default: 2 × median). */
-  gapThresholdMs?: number;
-};
-
-/**
- * Crosshair modes:
- * - `"normal"` — snaps only to the time axis (x), y follows the pointer (default)
- * - `"magnet"` — snaps x to the nearest bar and y to that bar's close
- * - `"magnetOHLC"` — snaps x to the nearest bar and y to the nearest of the bar's O/H/L/C values
- *   (within `snapThreshold` pixels; falls back to pointer y beyond the threshold)
- */
-export type CrosshairMode = "normal" | "magnet" | "magnetOHLC";
-
-export type CrosshairOptions = {
-  /** Snap mode (default: "normal") */
-  mode?: CrosshairMode;
-  /** Pixel distance within which OHLC snapping engages for `magnetOHLC` (default: 12) */
-  snapThreshold?: number;
-  /** Enable long-press crosshair lock on touch devices (default: true) */
-  lockOnLongPress?: boolean;
-};
-
-export type InteractionOptions = {
-  /**
-   * Enable the chart's own inertia tail after a wheel/trackpad gesture ends
-   * (default: true). Applies to both horizontal pan and zoom via the wheel.
-   *
-   * **Platform note:** on macOS, trackpads generate an OS-level momentum
-   * scroll that continues sending wheel events for a few hundred ms after
-   * the user lifts their fingers. Those are indistinguishable from user
-   * input and are always processed; this option only gates the synthetic
-   * inertia the chart adds on top once OS momentum has finished. On
-   * platforms without OS momentum (Windows mouse wheel, most Linux
-   * setups) the gesture stops immediately when this is false.
-   */
-  wheelInertia?: boolean;
-};
-
-/**
- * Hotkey actions. Either a drawing tool to activate (via Alt+letter),
- * `"cancel"` (Escape), or `"toggleOverlays"` (Ctrl+Alt+H — temporarily hide/show every series).
- */
-export type HotkeyAction = DrawingType | "cancel" | "toggleOverlays";
-
-/** Base chart rendering type */
-export type ChartType = "candlestick" | "line" | "mountain" | "ohlc";
-
-/** Data passed to the formatInfoOverlay callback */
-export type InfoOverlayData = {
-  candle: CandleData;
-  index: number;
-  paneId: string;
-  series: { label: string; color: string; value: unknown }[];
-};
-
-/** Duration presets for range selector */
-export type RangeDuration = "1D" | "1W" | "1M" | "3M" | "6M" | "YTD" | "1Y" | "ALL";
-
-// ============================================
-// Pane Configuration
-// ============================================
-
-export type ScaleMode = "linear" | "log" | "percent";
-
-export type PaneConfig = {
-  /** Unique pane identifier */
-  id: string;
-  /** Flex proportion for height allocation */
-  flex: number;
-  /** Y-axis scale mode for right scale (default: "linear") */
-  yScale?: ScaleMode;
-  /** Fixed Y-axis range for right scale (e.g., [0, 100] for RSI) */
-  yRange?: [number, number];
-  /** Horizontal reference lines (e.g., [30, 70] for RSI) */
-  referenceLines?: number[];
-  /** Reference line color */
-  referenceLineColor?: string;
-  /** Left scale configuration (enables dual-scale mode when present) */
-  leftScale?: {
-    mode?: ScaleMode;
-    range?: [number, number];
-    referenceLines?: number[];
-    referenceLineColor?: string;
-  };
-};
-
-export type LayoutConfig = {
-  /** Pane definitions */
-  panes: PaneConfig[];
-  /** Gap between panes in pixels (default: 4) */
-  gap?: number;
-  /** Show bottom scrollbar (default: true) */
-  scrollbar?: boolean;
-};
-
-// ============================================
-// Series Configuration
-// ============================================
-
-/** Built-in visual series types */
-export type BuiltinSeriesType =
-  | "line"
-  | "area"
-  | "histogram"
-  | "band"
-  | "cloud"
-  | "marker"
-  | "box"
-  | "heatmap";
-
-/** Visual series types (extensible via plugins) */
-export type SeriesType = BuiltinSeriesType | (string & {});
-
-export type SeriesConfig = {
-  /** Target pane: 'main' (overlay) or a specific pane id. Omit for auto-detection via __meta. */
-  pane?: "main" | string;
-  /** Scale assignment: 'right' (default) or 'left' for dual-scale panes */
-  scaleId?: "left" | "right";
-  /** Override auto-detected series type */
-  type?: SeriesType;
-  /** Primary color */
-  color?: string;
-  /** Line width in pixels (default: 1.5) */
-  lineWidth?: number;
-  /** Display label */
-  label?: string;
-  /** Initial visibility (default: true) */
-  visible?: boolean;
-  /** Max height as fraction of pane (0-1). Expands scale range so data occupies at most this ratio of pane height. Useful for volume overlay. */
-  maxHeightRatio?: number;
-  /** Fixed Y-axis range for the pane (e.g., [0, 100] for RSI). Applied when a new pane is created. */
-  yRange?: [number, number];
-  /** Horizontal reference lines (e.g., [30, 70] for RSI). Applied when a new pane is created. */
-  referenceLines?: number[];
-  /** Per-channel colors for multi-channel series (e.g., { upper: "#2196F3", lower: "#2196F3" } for bands) */
-  channelColors?: Record<string, string>;
-  /**
-   * Palette hint used by the auto-color rotation when `color` is not set.
-   * Populated from an indicator preset's `color` field when it's an array.
-   * Each new series using this same palette instance picks the next entry
-   * (wrapping), so stacking e.g. multiple SMAs produces distinct colors.
-   */
-  colorPalette?: readonly string[];
-};
-
-// ============================================
-// Series Handle (returned by addIndicator)
-// ============================================
-
-export type SeriesHandle = {
-  /** Unique series id */
-  readonly id: string;
-  /**
-   * Resolved series config (after preset defaults, palette rotation, and
-   * introspection). Useful for reading back the auto-assigned color.
-   */
-  readonly config: Readonly<SeriesConfig>;
-  /** Push a single data point (streaming). Accepts scalar or compound values. */
-  update(point: DataPoint<unknown>): void;
-  /** Replace all data */
-  setData<T>(data: DataPoint<T>[]): void;
-  /** Toggle visibility */
-  setVisible(visible: boolean): void;
-  /** Remove this series from the chart */
-  remove(): void;
-};
-
-// ============================================
-// Drawing Types
-// ============================================
-
-export type DrawingType =
-  | "hline"
-  | "trendline"
-  | "fibRetracement"
-  | "ray"
-  | "hray"
-  | "vline"
-  | "rectangle"
-  | "channel"
-  | "fibExtension"
-  | "textLabel"
-  | "arrow";
-
-export type DrawingBase = {
-  id: string;
-  type: DrawingType;
-  color?: string;
-  lineWidth?: number;
-};
-
-export type HLineDrawing = DrawingBase & {
-  type: "hline";
-  price: number;
-};
-
-export type TrendLineDrawing = DrawingBase & {
-  type: "trendline";
-  startTime: TimeValue;
-  startPrice: number;
-  endTime: TimeValue;
-  endPrice: number;
-};
-
-export type FibRetracementDrawing = DrawingBase & {
-  type: "fibRetracement";
-  startTime: TimeValue;
-  startPrice: number;
-  endTime: TimeValue;
-  endPrice: number;
-  levels?: number[];
-};
-
-export type RayDrawing = DrawingBase & {
-  type: "ray";
-  startTime: TimeValue;
-  startPrice: number;
-  endTime: TimeValue;
-  endPrice: number;
-};
-
-export type HRayDrawing = DrawingBase & {
-  type: "hray";
-  time: TimeValue;
-  price: number;
-};
-
-export type VLineDrawing = DrawingBase & {
-  type: "vline";
-  time: TimeValue;
-};
-
-export type RectangleDrawing = DrawingBase & {
-  type: "rectangle";
-  startTime: TimeValue;
-  startPrice: number;
-  endTime: TimeValue;
-  endPrice: number;
-  fillColor?: string;
-};
-
-export type ChannelDrawing = DrawingBase & {
-  type: "channel";
-  startTime: TimeValue;
-  startPrice: number;
-  endTime: TimeValue;
-  endPrice: number;
-  /** Price offset from the main line (positive = above) */
-  channelWidth: number;
-  fillColor?: string;
-};
-
-export type FibExtensionDrawing = DrawingBase & {
-  type: "fibExtension";
-  startTime: TimeValue;
-  startPrice: number;
-  endTime: TimeValue;
-  endPrice: number;
-  levels?: number[];
-};
-
-export type TextLabelDrawing = DrawingBase & {
-  type: "textLabel";
-  time: TimeValue;
-  price: number;
-  text: string;
-  fontSize?: number;
-  backgroundColor?: string;
-};
-
-export type ArrowDrawing = DrawingBase & {
-  type: "arrow";
-  startTime: TimeValue;
-  startPrice: number;
-  endTime: TimeValue;
-  endPrice: number;
-};
-
-export type Drawing =
-  | HLineDrawing
-  | TrendLineDrawing
-  | FibRetracementDrawing
-  | RayDrawing
-  | HRayDrawing
-  | VLineDrawing
-  | RectangleDrawing
-  | ChannelDrawing
-  | FibExtensionDrawing
-  | TextLabelDrawing
-  | ArrowDrawing;
-
-// ============================================
-// Multi-Timeframe Types
-// ============================================
-
-export type TimeframeOverlay = {
-  /** Identifier for this timeframe overlay */
-  id: string;
-  /** Higher timeframe candle data */
-  candles: CandleData[];
-  /** Timeframe label (e.g., '1W', '1M') */
-  timeframe: string;
-  /** Candle body color (default: semi-transparent) */
-  color?: string;
-  /** Opacity (default: 0.15) */
-  opacity?: number;
-};
-
-// ============================================
-// Backtest Result (compatible with trendcraft BacktestResult)
-// ============================================
-
-export type BacktestResultData = {
-  initialCapital: number;
-  finalCapital: number;
-  totalReturnPercent: number;
-  tradeCount: number;
-  winRate: number;
-  maxDrawdown: number;
-  sharpeRatio: number;
-  profitFactor: number;
-  trades: {
-    entryTime: number;
-    entryPrice: number;
-    exitTime: number;
-    exitPrice: number;
-    returnPercent: number;
-    direction?: string;
-    exitReason?: string;
-  }[];
-  drawdownPeriods: {
-    startTime: number;
-    troughTime: number;
-    recoveryTime?: number;
-    maxDepthPercent: number;
-    peakEquity: number;
-    troughEquity: number;
-  }[];
-};
-
-// ============================================
-// Pattern Signal (compatible with trendcraft PatternSignal)
-// ============================================
-
-export type ChartPatternSignal = {
-  time: number;
-  type: string;
-  pattern: {
-    startTime: number;
-    endTime: number;
-    keyPoints: { time: number; index: number; price: number; label: string }[];
-    neckline?: { startPrice: number; endPrice: number; slope: number; currentPrice: number };
-    target?: number;
-    stopLoss?: number;
-    height?: number;
-  };
-  confidence: number;
-  confirmed: boolean;
-};
-
-// ============================================
-// Signal / Trade overlay types
-// ============================================
-
-export type SignalMarker = {
-  time: TimeValue;
-  type: "buy" | "sell";
-  label?: string;
-};
-
-export type TradeMarker = {
-  entryTime: TimeValue;
-  entryPrice: number;
-  exitTime: TimeValue;
-  exitPrice: number;
-  direction?: "long" | "short";
-  returnPercent?: number;
-  exitReason?: string;
-};
-
-// ============================================
-// Event Types
-// ============================================
-
-export type ChartEvent =
-  | "crosshairMove"
-  | "visibleRangeChange"
-  | "click"
-  | "resize"
-  | "paneResize"
-  | "seriesAdded"
-  | "seriesRemoved"
-  | "dataFiltered"
-  | "drawingComplete"
-  | "error";
-
-export type CrosshairMoveData = {
-  time: TimeValue | null;
-  price: number | null;
-  x: number;
-  y: number;
-  paneId: string;
-};
-
-export type VisibleRangeChangeData = {
-  startTime: TimeValue;
-  endTime: TimeValue;
-  startIndex: number;
-  endIndex: number;
-};
-
-// ============================================
-// Chart Instance Interface
-// ============================================
-
-/** Summary info for a series */
-export type SeriesInfo = {
-  id: string;
-  paneId: string;
-  type: SeriesType;
-  label: string;
-  visible: boolean;
-};
-
-export type ChartInstance = {
-  // Data
-  setCandles(candles: CandleData[]): void;
-  updateCandle(candle: CandleData): void;
-  /** Batch multiple updates into a single render frame.
-   *  All mutations inside `fn` are deferred until the batch ends. */
-  batchUpdates(fn: () => void): void;
-
-  // Indicators (Series<T> native)
-  addIndicator<T>(series: DataPoint<T>[], config?: SeriesConfig): SeriesHandle;
-
-  // Series query
-  getAllSeries(): SeriesInfo[];
-  getVisibleRange(): VisibleRangeChangeData | null;
-
-  // Signals & Trades
-  addSignals(signals: SignalMarker[]): void;
-  addTrades(trades: TradeMarker[]): void;
-
-  // Drawings
-  addDrawing(drawing: Drawing): void;
-  removeDrawing(id: string): void;
-  getDrawings(): Drawing[];
-
-  // Drawing tool mode
-  setDrawingTool(tool: DrawingType | null): void;
-
-  // Multi-timeframe
-  addTimeframe(overlay: TimeframeOverlay): void;
-  removeTimeframe(id: string): void;
-
-  // Backtest visualization (trendcraft integration)
-  addBacktest(result: BacktestResultData): void;
-
-  // Pattern visualization
-  addPatterns(patterns: ChartPatternSignal[]): void;
-
-  // Score heatmap
-  addScores(scores: DataPoint<number | null>[]): void;
-
-  // Layout
-  setLayout(layout: LayoutConfig): void;
-
-  // Viewport
-  setVisibleRange(start: TimeValue, end: TimeValue): void;
-  setVisibleRangeByDuration(duration: RangeDuration): void;
-  fitContent(): void;
-
-  /**
-   * Programmatically set or clear the crosshair position by time value.
-   * Used by `syncCharts()` to mirror the crosshair across multiple chart instances.
-   * Pass `null` to hide the crosshair.
-   */
-  setCrosshair(time: TimeValue | null): void;
-
-  // Events
-  on<E extends ChartEvent>(event: E, handler: (data: unknown) => void): void;
-  off<E extends ChartEvent>(event: E, handler: (data: unknown) => void): void;
-
-  // Theme & Chart Type
-  setTheme(theme: "dark" | "light" | ThemeColors): void;
-  setChartType(type: ChartType): void;
-
-  // Volume pane visibility
-  /** Show or hide the volume pane */
-  setShowVolume(show: boolean): void;
-
-  /**
-   * Apply a partial options update at runtime.
-   *
-   * Accepts the same shape as `createChart`'s options, but applies only the
-   * provided fields. Use this from reactive wrappers to propagate option
-   * changes after chart creation. Fields that cannot be changed at runtime
-   * (e.g. `pixelRatio`, `fontFamily`, `scrollSensitivity`, `locale`,
-   * `formatInfoOverlay`) emit a warning via the `error` event and are ignored.
-   */
-  applyOptions(options: Partial<ChartOptions>): void;
-
-  // Plugins
-  /** Register a custom series renderer plugin */
-  registerRenderer<TConfig>(plugin: import("./plugin-types").SeriesRendererPlugin<TConfig>): void;
-  /** Register a pane primitive plugin */
-  registerPrimitive<TState>(plugin: import("./plugin-types").PrimitivePlugin<TState>): void;
-  /** Remove a primitive by name */
-  removePrimitive(name: string): void;
-
-  // Extensibility
-  /** Register a custom introspection rule for shape detection */
-  addRule(rule: import("./series-registry").IntrospectionRule): void;
-  /** Register a visual preset (colors, lineWidth, etc.) for a rule name */
-  addPreset(name: string, preset: import("../integration/indicator-presets").IndicatorPreset): void;
-
-  // Export
-  toImage(type?: string, quality?: number, timeoutMs?: number): Promise<Blob>;
-
-  // Lifecycle
-  resize(width: number, height: number): void;
-  destroy(): void;
-};
-
-// ============================================
-// Internal Rendering Types
-// ============================================
-
-/** Computed pane dimensions (from layout engine) */
-export type PaneRect = {
-  id: string;
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-  config: PaneConfig;
-};
-
-/** Computed series rendering info */
-export type ResolvedSeries = {
-  id: string;
-  paneId: string;
-  type: SeriesType;
-  config: SeriesConfig;
-  data: DataPoint<unknown>[];
-  /** Decomposed numeric channels for compound types */
-  channels: Map<string, (number | null)[]>;
-};
+export * from "./types/fundamental";
+export * from "./types/theme";
+export * from "./types/drawing";
+export * from "./types/series";
+export * from "./types/pane";
+export * from "./types/config";
+export * from "./types/event";
+export * from "./types/integration";
+export * from "./types/chart-instance";
